updates for various bugs--part1
1. Intro: from second level page, the Android back button should take us
back to the first level page of the Intro.
2. Signup/Signin forms: don’t hide the hint in the email field until the
user actually begins typing. They need to know they should be typing an
email address
3. Homescreen should be refresh-able.  We want to make sure that if the
user has created a new item recently, or someone else has created an
item and they are talking about it on somne channel (chat…) they can see
it in the homescreen/Feed. A “pulldown-gesture” to refresh is our
preferred approach